+++ NEWS.md
@@ -1,3 +1,5 @@
+# routr (development version)
+
 # routr 1.0.0
 
 * Use native rlang type checking instead of assertthat