+++ NEWS.md
@@ -14,6 +14,7 @@
   All versions of the report (with regard to parameters) are deleted. If you use
   a format-specific path only cached versions of that format is deleted (#30)
 * Add support for OpenTelemetry through the otel package (#33)
+* The `openapi_route()` now better handles docs served through a proxy
 
 # routr 1.0.0